refactor(coin_asset): replacing GenerateCoinAssetID with getCoinAssetID
