+++ utilities/rest/id_getters/docs/split.go
@@ -3,7 +3,7 @@ package docs
 import (
 	"net/http"
 
-	"github.com/AssetMantle/schema/go/documents/base"
+	baseDocuments "github.com/AssetMantle/schema/go/documents/base"
 	errorConstants "github.com/AssetMantle/schema/go/errors/constants"
 	"github.com/AssetMantle/schema/go/ids"
 	baseIDs "github.com/AssetMantle/schema/go/ids/base"
@@ -31,7 +31,7 @@ func splitIDHandler(context client.Context) http.HandlerFunc {
 
 		var coinID ids.AssetID
 		for _, coin := range coins {
-			coinID = base.GenerateCoinAssetID(coin.Denom)
+			coinID = baseDocuments.NewCoinAsset(coin.Denom).GetCoinAssetID()
 		}
 
 		rest.PostProcessResponse(responseWriter, context, newResponse(baseIDs.NewSplitID(coinID, fromID.(ids.IdentityID)).AsString(), "", nil))
